Connect DataCollectionModal
