+++ app/components/Views/DataCollectionModal/styles.ts
@@ -12,6 +12,13 @@ const createStyles = () =>
       display: 'flex',
       flexDirection: 'column',
     },
+    wrapper: {
+      display: 'flex',
+      flexDirection: 'column',
+      alignItems: 'center',
+      marginTop: 10,
+      marginBottom: 10,
+    },
   });
 
 export default createStyles;
