+++ app/components/Views/ExperienceEnhancerModal/index.tsx
@@ -1,8 +1,10 @@
-import React from 'react';
+import React, { useRef } from 'react';
 import { View } from 'react-native';
 
 import { strings } from '../../../../locales/i18n';
-import BottomSheet from '../../../component-library/components/BottomSheets/BottomSheet';
+import BottomSheet, {
+  BottomSheetRef,
+} from '../../../component-library/components/BottomSheets/BottomSheet';
 import Text, {
   TextVariant,
 } from '../../../component-library/components/Texts/Text';
@@ -19,6 +21,7 @@ import { setDataCollectionForMarketing } from '../../../actions/security';
 
 const ExperienceEnhancerModal = () => {
   const styles = createStyles();
+  const bottomSheetRef = useRef<BottomSheetRef>(null);
 
   const cancelButtonProps: ButtonProps = {
     variant: ButtonVariants.Secondary,
@@ -26,6 +29,7 @@ const ExperienceEnhancerModal = () => {
     size: ButtonSize.Lg,
     onPress: () => {
       setDataCollectionForMarketing(false);
+      bottomSheetRef.current?.onCloseBottomSheet();
     },
   };
 
@@ -35,11 +39,12 @@ const ExperienceEnhancerModal = () => {
     size: ButtonSize.Lg,
     onPress: () => {
       setDataCollectionForMarketing(true);
+      bottomSheetRef.current?.onCloseBottomSheet();
     },
   };
 
   return (
-    <BottomSheet>
+    <BottomSheet ref={bottomSheetRef}>
       <Text variant={TextVariant.HeadingMD} style={styles.title}>
         {strings('experience_enhancer_modal.title')}
       </Text>
